playground/examples/webapp/requirements.txt
flask==3.0.0
gunicorn==21.0.0
psycopg2-binary==2.9.9
redis==5.0.0
prometheus_client==0.19.0
